Renamed references to Pokemon, added Equipment class
